inconsistency fixed... residual ms bug in calendar class led to 3999 instead of 4000, e.g.

+++ chargecarprize/javatestsuite/ChargeCarPrize/src/chargecar/policies/NaiveBufferPolicy.java
@@ -3,8 +3,6 @@
  */
 package chargecar.policies;
 
-import java.math.BigDecimal;
-
 import chargecar.battery.BatteryModel;
 import chargecar.capacitor.CapacitorModel;
 import chargecar.util.PointFeatures;
@@ -32,7 +30,7 @@ public class NaiveBufferPolicy implements Policy{
 	
 	@Override
 	public PowerFlows calculatePowerFlows(PointFeatures pf) {
-		double watts = pf.getPowerDemand().doubleValue();
+		double watts = pf.getPowerDemand();
 		int periodMS = pf.getPeriodMS();
 		double min = modelCap.getMinCurrent(periodMS);
 		double max = modelCap.getMaxCurrent(periodMS);
